requirements.txt
charset-normalizer 3.4.4
idna               3.11
pip                24.0
python-dotenv      1.2.1
requests           2.32.5
urllib3            2.5.0